strlen uses sizeof not int

+++ ft_strdup.c
@@ -6,13 +6,13 @@
 /*   By: jfritz <jfritz@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/18 12:15:21 by jfritz            #+#    #+#             */
-/*   Updated: 2021/06/18 15:06:29 by jfritz           ###   ########.fr       */
+/*   Updated: 2021/06/18 15:08:14 by jfritz           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *c);
+size_t	ft_strlen(const char *c);
 
 char	*ft_strcpy(char *dest, char *src)
 {
